Restructured files under /src/api/v1/products and repaired paths accordingly
