+++ src/tests/api/v1/products/add/getProductsByCategoryAndPrice.test.js
@@ -1,4 +1,4 @@
-const getProductsByCategoryAndPrice = require('../../../../routes/api/v1/products/getProductsByCategoryAndPrice');
+const getProductsByCategoryAndPrice = require('../../../../../routes/api/v1/products/add/getProductsByCategoryAndPrice');
 
 describe('Tests for getting product details by category and price from external API', () => {
   const validProductParameters = {
